refactor(chat-server): extract state, push, and ws into chat-server/
Splits chat-server.ts's three remaining biggest cohesive concerns
into per-purpose modules. No behavior change; chat-server.ts re-exports
broadcast / setOnNewMessage / clearOnNewMessage / setAgentPresence /
setOnGroupUpdated / clearOnGroupUpdated / broadcastRooms so existing
imports from channels/local-chat.ts and src/index.ts keep working.

- chat-server/state.ts (153 lines):
    in-memory client registry, broadcast (with WS fanout + Web Push),
    member presence, channel adapter callback hooks
- chat-server/push.ts (115 lines):
    VAPID init, isValidPushEndpoint allowlist, sendPushForMessage
    (called from state.broadcast on user-typed messages)
- chat-server/ws.ts (249 lines):
    setupWebSocket — keepalive timer, /ws upgrade auth, AUTH/JOIN/
    TYPING/MESSAGE/DELETE_MESSAGE message handlers
- chat-server.ts: 1581 -> 1104 lines. Now contains the http handler
  with REST routes, file-upload helpers, static PWA serving,
  startChatServer/stopChatServer, and a thin re-export shim.

Also cleaned orphan imports left behind: WebSocketServer, WebSocket,
deleteChatMessage, getChatAgentToken, deletePushSubscription,
getPushSubscriptionsExcludingIdentity, webPush, ASSISTANT_NAME.
